requirements.txt
Flask==3.1.1
twilio==9.6.5
python-dotenv==1.1.1
openai==1.97.0